feat(dynamic-key): first proof of concept for user defined dynamic `localStorageKey` in injected script
This proof of concept injects the fouc killer script during the `page` stage of rendering. This means that fouc is not eliminated because the script is run after the initial CSS is loaded.

However, this proof of concept demonstrates using virtual modules to dynamically set a user defined `localStorageKey` by importing it into a script, which is then injected into every page.

The first working version of this integration injects a script with a static local storage key using `head-inline`, this ensures the script is run before the intiial CSS is loaded on the page.

However, `head-inline` scripts are not processed or resolved by Vite, so dynamically setting the `localStorageKey` is challenging.

In this proof of concept, the dynamic `localStorageKey` is set by importing a virtual config module with the user defined key into a separate `foucKillerScript` file. This file is then resolved with `createResolver` and passed to the `injectScript` to be injected as a processed and resolved module at the `page` stage, after the initial CSS is loaded. This is why this proof of concept does not eliminate fouc, `head-inline` scripts are not processed by Vite so `page` is the only stage that can be used for dynamic script injection.