fix(dynamic-key): Eliminates FOUC while allowing a user defined `localStorageKey`
This implementaions splits the script into two parts: an inline script and a processed script.

The inline script is defined directly within the `injectScript` function using `head-inline` and is a simple script that gets the value of the user defined, or default, `localStorageKey` and sets the theme based on it.

The second part of the script uses the `page` stage so it can be defined in a different file and import the virtual config module. This way, it too can utilize the user defined, or default, `localStorageKey` values; It then listens for changes in local storage and toggles dark class accordingly.